Correcting slurm restart logic
Fixing Slurm restart in Master and compute nodes. With
respect to compute node restart, we are checking path existence
in master node which is wrong, now moved this logic to command line
so it executes in compute node. And in Master slurm restart we need to pass
restart command as list.

Signed-off-by: Mohan Gandhi <mohgan@amazon.com>