Wake-up Torque scheduler when attaching new nodes
This fix wants to address a problem with the Torque scheduler in the
scenario when the cluster has no compute nodes available.
Jobs get scheduled and go in Q state, but even when new compute nodes
become available to the cluster, the jobs remain in Q state.
This has a major impact to customers that see the cluster scaling up to
the max size but without any job running.

From tested approaches and according to Adaptive, one way to recover
from this situation is to reset the Torque server flag 'scheduling=true'
This works when the node fully joined the cluster and it's 'free'.

This patch implements this approach, when a node is added to the
cluster and becomes free, the 'scheduling' flag is reset.

Torque server restart commands have been removed because according to
official documentation (https://tinyurl.com/y7vjrztw), "When you make
changes to nodes by directly editing the nodes file, you must restart
pbs_server for those changes to take effect. **Changes made using qmgr
do not require a restart.**"
Restart the server can impact the user's experience for a bunch of
seconds, moreover frequent restarts when adding and removing nodes may
undermine the stability of the pbs_server and pbs_sched processes.

Signed-off-by: Maurizio Melato mmelato@amazon.com