Adding slots information to slurm
On host addition, the slurm configuration is changed to add the number
of slots for the compute node. This information is coming from the sqs
message sent by the compute node itself.

In order to allow to submit more than one job per node, the scheduler
needs also to be configured to use CPUs as consumable resource. This is
done setting the params SelectType and SelectTypeParameters as described
into https://slurm.schedmd.com/slurm.conf.html

Signed-off-by: Luca Carrogu <carrogu@amazon.com>