sqswatcher: avoid to process multiple times a bad host
Before this, if the hostname was empty or when an exception was raised
the related message was not removed from the queue.

The sqswatcher process crashed with an exception and a new sqswatcher process started.
The bad message remained in the queue and after the SQS visibility timeout
it was processed by the new sqswatcher process, crashing again.

Now we are logging the error, breaking the loop and deleting the message from the queue.
If a bad host is not added to the cluster,
it will be be automatically stopped after the scaledown_idletime.

Signed-off-by: Enrico Usai <usai@amazon.com>