nodewatcher: fix jobs detection for the sge scheduler
When the host was not added correctly to the cluster,
the command idle-nodes returned with exit code 0 and with the following message:
"error: denied: host xxx is neither submit nor admin host"

The wrong logic erroneously returned _jobs=True and the host was never
removed from the cluster. Now we are checking for error messages and
assuming no active jobs in the current host.

I also verified Torque and SLURM and the logic is correct.

Signed-off-by: Enrico Usai <usai@amazon.com>