Requeue instance terminate message if host is not removed from the scheduler
Requeue the instance terminate sqs message if one of the commands to remove
the host from the scheduler fail, in order to avoid the "dead node"
situation in SGE.

A "dead node" occurs when the instance is terminated but it's not
deleted from the scheduler configuration, because a job still references
it.

With this patch, if one of the commands to remove the host from the
scheduler configuration fail, then the sqs message containing the
instance terminate message is requeued in order to be reprocessed later.

When all the jobs referencing the terminated instance will be delete,
 then the host will be removed from the scheduler.

Signed-off-by: Luca Carrogu <carrogu@amazon.com>