slurm: add other reasons for the cluster scale up
We were scaling up only when the job was waiting for resources
to become available (Job reason code = Resources).

In case of a requeued job the cluster didn't scale up because the
pending reason is NodeDown and then BeginTime.
I'm adding them to the list of reason for the scale up.

I also added Priority because I think that we can scale up if the job
is waiting for other higher priority jobs.

Signed-off-by: Enrico Usai <usai@amazon.com>