jobwatcher: re-compute instance properties at every iteration to support updates
Instance properties containing the number of vcpus to use whan adding a node
to the scheduler was computed only when the jobwatcher is initialized and was
using a static instance type value read from the jobwatcher config file.
In case of an update of the compute instance type the jobwatcher was still
using the old cpu values to compute the number of instances to start when
scaling up the cluster.
Now the instance properties are re-computed at every iteration of the jobwatcher
and the instance type is retrieved from the cloudformation stack parameters.
A better solution would consist in updating and reloading the node daemons
every time a pcluster update operation is performed.

Signed-off-by: Francesco De Martino <fdm@amazon.com>