jobwatcher: do not scale-up if CPU contraints cannot be satisfied
Currently with Slurm and Torque when a job is submitted and the requested
slots per node are bigger that the number of available vcpus on a single
instance the jobwatcher is selecting a bigger number of nodes to fit the
requested CPUs. This operation/assumption is incorrect since the scheduler
will never schdule the job due to the fact that the CPU contraints are never
satisfied. This causes the cluster to spin up new capacity that will stay
idle forever or until the job is canceled.
This commit fixes the above mentioned issue.

Signed-off-by: Francesco De Martino <fdm@amazon.com>