Do not count as busy nodes locked by the nodewatcher
after a node is locked by the nodewatcher it soon gets terminated
and the ASG desired gets decremented. For this reason jobwatcher
should not consider such nodes as busy since they are already removed
from the ASG desired count. In case we consider them as busy the cluster
is going to overscale.

Signed-off-by: Francesco De Martino <fdm@amazon.com>