Manage possible exceptions from batch hosts addition or removal
Prior the patch the system was unable to recover if one or more host couldn't be added to the scheduler configuration.
The symptoms were that the SQS messages were not re-queued with a decreasing TTL, and the total number of messages grew indefinitely, causing the impossibility of adding new nodes.
Logic was already present for SLURM plugin, moved in common place to be used also by SGE and Torque declination.

This fix https://github.com/aws/aws-parallelcluster/issues/1327

Signed-off-by: Luca Carrogu <carrogu@amazon.com>