slurm: do not count down nodes as busy
In case a node is down and it's not in ASG this causes overscaling.
Note that this condition is not triggered by nodes that are set to DRAIN
by nodewatcher before being terminated. These nodes will preserve the
drain state despite being unreachable (drain*).

The downside to this is that when a node is unhealthy, hence set
to DOWN by the scheduler, its replacement is launched only after
the node is terminated and removed from the scheduler, which means
waiting for the SlurmdTimeout+nodewatcher timeout+processing of instance
terminate event.

Signed-off-by: Francesco De Martino <fdm@amazon.com>