sqswatcher: prevent message retries from populating slurm nodes config with zombie nodes
Due to the retries on failed SQS events, under certain circumstances the
retry on a ADD event can be processed after a successful REMOVE event for
the same instance. This causes terminated instance types to be added back to
slurm config and never removed.
With this change retries on failed ADD events are dropped and, in case of
colliding messages for the same hostname, REMOVE events are always prevailing
over ADD ones.

Signed-off-by: Francesco De Martino <fdm@amazon.com>