jobwatcher: retrieve busy nodes before computing required ones
If jobs are scheduled to nodes between the moment the jobwatcher
computes the number of requried nodes (based on the pending jobs) and
when the list of busy nodes is fetched then the cluster over-scales.
Moving get_busy_nodes before get_required_nodes in order to prevent
this from happening.

Signed-off-by: Francesco De Martino <fdm@amazon.com>