Don't log insufficient node resources message
Avoid generating a log message that states an existing node does not
have sufficient resources (e.g., slots) to run a job. If a message were
logged for this case, it would lead to log pollution in the case of a
cluster with many compute nodes and many pending jobs that could not be
placed on any of those nodes.

Signed-off-by: Tim Lane <tilne@amazon.com>