Configure hostname and DNS cluster domain by using Route53 hosted zone
Within the cluster we're creating a private hosted zone named <clustername>.parallelcluster

With this patch we're extending DNS resolver configuration
by appending the Route53 domain created within the cluster.

The private domain will be added as a "search" domain in the resolv.conf file
by dhclient in Centos6, CentOS7, Alinux, Alinux2 and Ubuntu16
and by systemd-resolved in Ubuntu18.
In this way we'll be able to resolve the short version of the FQDN.

In the HIT case we're also configuring the hostname to be something in the form
hostname: $QUEUE-static-$INSTANCE_TYPE_1-[1-$MIN1]
NOTE: we're replacing the "." in the instance type with the "-" character
because the period is used in hostnames to split subdomains.

OS specific settings:
* I added the bind-utils package to Centos7 to have the "dig" utility,
  useful to check DNS resolution.
* In Centos6 there is a dhclient configuration file for eth0 so we're
  setting the domain tin the /etc/dhcp/dhclient-eth0.conf file too.

Other changes:
+ Refactored the code to configure the hostname to be a function.
+ Moved cloudwatch_agent config (and install) in prep_env to
  have dns_config (and pre-install) logs available on cloudwatch.
+ Moved _setup_python recipe in prep_env to have the python virtual env available
  for dns_config and cloudwatch_agent install actions.
+ Added hosted_zone configuration parameter to clustermgtd.conf and slurm_resume.conf

References:
* chef two-pass model: https://coderanger.net/two-pass/
* shell_out: https://docs.chef.io/dsl_recipe/#shell_out-1
* lazy: https://docs.chef.io/resource_common/#lazy-evaluation
* replace_or_add: https://github.com/sous-chefs/line/blob/master/documentation/resources/replace_or_add.md
* dhclient: https://wiki.archlinux.org/index.php/NetworkManager#DHCP_client
* systemd-resolved: https://wiki.archlinux.org/index.php/Systemd-resolved
* dynamodb get-item: https://docs.aws.amazon.com/cli/latest/reference/dynamodb/get-item.html

Signed-off-by: Enrico Usai <usai@amazon.com>