Upload cluster metrics to CloudWatch for slurm
With this commit, cluster metrics are uploaded to CloudWatch. In order not to impact the job of the current clustermgtd, a separate thread was created to upload those metrics. Threads were used instead of processes as the functions here are I/O bound, threads are thus suitable. Some code has been added to the current deamon so that it saves the information it requests from the scheduler or from ec2. This information is then used by the second thread to upload the cluster metrics.

Signed-off-by: Alexandre Gobeaux <gobeaa@amazon.com>