Fix Exceptions handling in threads
The previous commit for handing thread exceptions was a wrong. This commit takes a new approach: instead of making the two deamons fail if one fails, we try to restart the failing deamon if possible. If an error happens while restarting, the main function is stopped and will be launched again thanks to the retry. Testing was done manually.

Signed-off-by: Alexandre Gobeaux <gobeaa@amazon.com>