Handle log.exception differently
Before this commit, custom exceptions were created and used to make the difference between the threads that needed to be restarted. We decided to use an alternative way that is more compact (since we log exceptions differently, we do not need the custom child threads exceptions anymore). Testing was done manually, trying to raise an exception in the two children deamons to verify that the trace was logged.

Signed-off-by: Alexandre Gobeaux <gobeaa@amazon.com>