Force nodes to down and power_save when stopping the cluster
If a cluster is stopped while a node is powering-up (alloc#-idle#),
node is kept in the powering-up state on cluster start.

This makes the node unavailable for the entire ResumeTimeout which is 60 minutes.
Slurm is ignoring the transition to power_down if we don't put the node to down first.

Copy Changelog from 2.11 branch.

Signed-off-by: Enrico Usai <usai@amazon.com>