Force nodes to down and power_save when stopping the cluster
If a cluster is stopped while a node is powering-up (alloc#-idle#),
node is kept in the powering-up state on cluster start.

This makes the node unavailable for the entire ResumeTimeout which is 60 minutes.
Slurm is ignoring the transition to power_down if we don't put the node to down first.

From @demartinofra

## Manual test
* Created a cluster and submitted a job on it
* When the node was powering up stopped the cluster and verified the node is correctly marked as power down
* Restarted the cluster and verified the node is back to powering save state (after about 2 minutes)
* Job ran correctly in the new node.

Signed-off-by: Enrico Usai <usai@amazon.com>