Improve log message of smart capacity failover
Only log compute resources name instead of node names when detect or reset insuffcient capacity nodes. Don't treat insufficient capacity nodes as unhealthy nodes to not print each ice node message in the logs. The reason is that there may be lots of dynamic nodes in a compute resources. Log all the nodes belong to a compute resource will be a lot.

Signed-off-by: chenwany <chenwany@amazon.com>