Handle case when instances don't have all EC2 info yet
Handle corner case when instance is just launched and the describe doesn't report yet all the EC2 info like the private IP address.
In this case, the get_cluster_instances must not completely fail, but only skip the instance for which there aren't all the info yet.

The get_cluster_instances is used in the compute fleet stopped loop to keep the nodes to down and in the manage cluster loop to update slurm nodes with EC2 info and to terminate orphaned nodes.

Signed-off-by: Luca Carrogu <carrogu@amazon.com>