[SlurmPlugin] Log exceptions raised when reading json
While reading json files such as `run_instances_overrides.json`, the read operation may fail due to use of non-standard characters. This is difficult to debug since at the moment we use a default value when there is an error and furthermore the error is not logged.
This commit logs the exception in situations where we fallback to a default value.