Use CreateFleet API when multiple instance types are specified in the compute resource configuration
This patch permits to call boto3 `create-fleet` API when `InstanceTypeList` is specified in the cluster configuration.
The create-fleet response just contain the instance-id list, so we need to execute an additional describe-instances call
to retrieve other information like ip address to be saved in DynamoDB.
Raises `Ec2InstancesInfoUnavailable` exception in `Ec2CreateFleetManager` too to signal the case when describe-instances call
returns incomplete info for 3 times (15 seconds).

To pass the queue/compute-resource configuration parameters to the create-fleet api I'm using the `cluster-config.yaml` file.
This file contains the settings from the user and default values (e.g. `AllocationStrategy`).

The create-fleet api can return a list of Errors,
they will be logged as errors if there are no returned instances, as warning otherwise.

## New class hierarchy: `*FleetManager`
Within this patch we're moving responsibility to launch the EC2 instances from `InstanceManager` to `*FleetManager`.
The fleet manager will be selected (`Ec2RunInstancesManager` - `run-instances` API vs `Ec2CreateFleetManager` - `create-fleet` API)
according to the configuration of the given compute resource.

## Run-instances-overrides
This patch renames internal variables from `run_instances_overrides` to `launch_overrides` to clarify
the code is able to pass the overridden parameters to both the APIs.
The configuration parameter and the file has not been renamed to preserve back-word compatibility.

## All or nothing
I'm adding a basic support for All-or-nothing feature, setting `MinTargetCapacity` to 1
when all_or_nothing is not configured.
If the minimum target capacity is not reached, the fleet launches no instances,
this might differ from run-instances approach where the run-instances call fails, to be verified.

## Multi-AZ support
In the create-fleet API we're explicitly setting `SingleAvailabilityZone` to `True`, the multi-AZ support is not included.

## Tests
Manually tested by updating the node package in a running cluster and reloading clustermgtd daemon
(slurm_resume changes don't require any reload).
To test this package I had to manually modify the `cluster-config.yaml` by adding `AllocationStrategy` and `InstanceTypeList` params
and adding some required policies to the Role attached to the head node.
Required policies are: `ec2:CreateFleet`, `ec2:RunInstances`, `ec2:DescribeCapacityReservations` and `ec2:CreateTags`.

### Unit tests
* Move test_instance_manager file outside slurm_resources folder.
* Move tests from test_instance_manager to test_fleet_manager.
* Add unit tests for *FleetManager classes, to verify create_fleet and describe_instances boto3 calls
* Mock launch_instances from fleet manager rather than boto3 calls in clustermgtd and resume unit tests.

## References
* https://boto3.amazonaws.com/v1/documentation/api/latest/reference/services/ec2.html#EC2.Client.create_fleet

Signed-off-by: Enrico Usai <usai@amazon.com>