Distinguish between run-instances and create-fleet overrides json files
The two files, expected by both clustermgtd and slurm_resume daemons are:
* `/opt/slurm/etc/pcluster/run_instances_overrides.json`
* `/opt/slurm/etc/pcluster/create_fleet_overrides.json`

They will be passed to the fleet manager instance according to the API
to use for the given compute resource.

I modified the `read_json` file to avoid printing an INFO message with "exception" word
into it, when the file doesn't exist, to avoid to confuse users with this Errno/exception word.
Before of this patch we had:
```
2022-09-14 09:42:24,675 - [common.utils:read_json] - INFO - Unable to read file
'/opt/slurm/etc/pcluster/run_instances_overrides.json'
due to an exception: [Errno 2] No such file or directory:
'/opt/slurm/etc/pcluster/run_instances_overrides.json'. Using default: {}
````
Now we'll have this INFO message only when we're not able to parse the file (e.g. `JSONDecodeError`).`

### Unit tests changes

I extended and fixed `read_json` unit tests that weren't using the files in the right folder
and they were always failing because of `FileNotFoundError`.

I added an empty override dict {} to all the read_json side effect,
because right now we are parsing 3 files rather than 2.
I also removed useless overrides content when not needed, replacing the values with {}

Signed-off-by: Enrico Usai <usai@amazon.com>