Add unit test scenario we currently cannot handle
The added unit test scenario is legit for Slurm's syntax, but
node's batching logic of input parameters for scontrol update
cannot handle strings using different grouping.

This situation should be impossible in the current logic of
clustermgtd, but in the future we should try to support any
legit combination of arguments for scontrol update.

Signed-off-by: Jacopo De Amicis <jdamicis@amazon.it>