Do not drain nodes failing EC2 checks after scontrol reboot
Add a new timeout that controls how long after the booting of slurmd
(which happens at the first start of a compute node or after
an scontrol reboot operation) clustermgtd is allowed to drain a compute
node if the corresponding EC2 instance has the EC2 status checks failing.

After performing scontrol reboot, EC2 might signal the instance as unhealthy
for a certain amount of time, even after the instance has fully booted and
the Slurm compute node has registered itself to the Slurm control daemon. This
is just an issue of timing in updating the status of the instance on the EC2
side.

This is related to https://github.com/aws/aws-parallelcluster/issues/4751

Add some logging about nodes failing EC2 health checks that were not set
to DRAIN because they had their Slurmd daemon recently restarted.

Add unit test case for time_is_up() to cover case when initial_time is None.

Add python typing hint to time_is_up()

Signed-off-by: Jacopo De Amicis <jdamicis@amazon.it>