Allow clustermgtd to handle compute nodes rebooted via `scontrol reboot` (#490)
This is related to https://github.com/aws/aws-parallelcluster/issues/4751

Do not consider nodes failing EC2 health checks while they are in
REBOOT_ISSUED state.

Do not consider nodes failing EC2 health check for a certain time after 
slurmd has started on the compute nodes.

Add a new timeout that controls how long after the booting of slurmd
(which happens at the first start of a compute node or after
an scontrol reboot operation) clustermgtd is allowed to drain a compute
node if the corresponding EC2 instance has the EC2 status checks failing.

After performing scontrol reboot, EC2 might signal the instance as unhealthy
for a certain amount of time, even after the instance has fully booted and
the Slurm compute node has registered itself to the Slurm control daemon. This
is just an issue of timing in updating the status of the instance on the EC2
side.

Add slurmdstarttime attribute to SlurmNode class to track the SlurmdStartTime
field returned by scontrol show nodes, which is the instant at which the Slurmd
daemon started on the compute node.

Fix small bug that could cause an inconsistent stripping of spaces
across the fields extracted from scontrol show nodes

Fix scontrol show nodes parser handle "NextState=" when searching
for "State=". When rebooting nodes via scontrol reboot, scontrol show node 
shows a field "NextState=". The previous parsing syntax caught that as
a second instance of "State=".

Add unit tests to cover the modifications introduced in this PR.

Add entry to the CHANGELOG.md

Signed-off-by: Jacopo De Amicis <jdamicis@amazon.it>