Exclude nodes in maintenance from static nodes to be replaced
After this patch the nodes in `MAINTENANCE+RESERVED` state
will be excluded by the static nodes in replacement list.

Will see in the "replacement list" only nodes that are not yet up (as before)
and nodes that are not in maintenance.

This mechanism permits the daemons (or the user) to avoid replacement
of turned down static nodes (`DOWN`) by putting them in maintenance as preliminary step.

This works only for nodes in both `MAINTENANCE` and `RESERVED` state,
nodes only in `RESERVED` state will be replaced as before.

In the future, we can extend this mechanism to have an additional field to check
(e.g. skip them only if the maintenance is set by the `root` user).

Signed-off-by: Enrico Usai <usai@amazon.com>