Raise exception when CreateFleet doesn't return any instance
Raise an exception when CreateFleet doesn't return any instance and the error list in the CreateFleet contains only one error entry.
The exception raised is built with the same error code coming from the CreateFleet response, so that this error code can be set into the reason when putting the Slurm nodes into DOWN.
This is useful to avoid to trigger the fast capacity failover (error code InsufficientInstanceCapacity) when the CreateFleet call doesn't return any instance because of throttling (error code RequestLimitExceeded).

Signed-off-by: Luca Carrogu <carrogu@amazon.com>