Introduce CapacityBlockManager to handle instances part of a Capacity Block reservation
### Capacity Block reservation

Capacity Block reservation for ML (CB) allow AWS users to reserve blocks of GPU capacity for a future start date
and for a time duration of their choice.
They are capacity reservations of type `capacity-block`, with additional attributes like:
`StartDate`, `EndDate` and additional states.

Nodes part of the CB cannot be used until the CB time started.
Jobs assigned to these nodes will be kept in pending until the CB is active,
to achieve this we're going to use Slurm reservation, in particular we'll set the nodes
in maintenance state when CB is not yet active or when CB time is expired.

### CapacityBlockManager

CapacityBlockManager is a new class that will perform the following actions:
* when CB is not active or expired, a new Slurm reservation will be created/updated
* when CB is active, Slurm reservation will be removed and nodes will become standard static instances

Slurm reservation name will be "pcluster-{capacity_block_reservation_id}"

CapacityBlockManager will be initialized by clustermgtd with region, boto3_config and fleet_config info.
Fleet config is used to retrieve capacity reservation ids and type (ondemand vs capacity-block).
The manager will reload fleet-config info and capacity reservation info every time the daemon is restarted
or when the config is modified.
The logic updates capacity block reservation info every 10 minutes.

The manager will remove the nodes associated with capacity-block reservations that are not active
from the list of unhealty static nodenames.

### CapacityBlock

CapacityBlock is a new class to store internal info about the capacity block,
merging them from ec2 (e.g. capacity block state) and from the config (i.e. list of nodes associated to it)

### Leftover slurm reservations

When a CB is removed from a cluster config during a cluster update
we need to remove related slurm reservations.
We're checking slurm reservations for the given nodes and deleting them
when are managed by CapacityBlockManager but no longer in the config.

Signed-off-by: Enrico Usai <usai@amazon.com>