Cleanup all the leftover slurm reservations even if not related to given nodes
Previously the code was only deleting leftover slurm reservations
from the given nodes, but if the CB is moved from a queue to another
the list of nodes might not contain all the reservation.

With this new logic we're retrieving all the slurm reservations from slurm
and deleting them if they are no longer associated with existing CBs in the config.

Signed-off-by: Enrico Usai <usai@amazon.com>