Mark nodes as reserved only if Slurm reservation is correctly created
Previously the node was considered as reserved and removed from unhealthy nodes,
even if the slurm reservation process was failing.
Now we're checking if the slurm reservation has been correctly created/updated.

If the slurm reservation actions for ALL the CBs are failing,
we're not marking the nodes as reserved and not updating
the internal list of capacity blocks and update time.

If one of the reservation is updated correctly
the list of reserved nodes and update time attributes are updated.

Extended unit tests accordingly.

Signed-off-by: Enrico Usai <usai@amazon.com>