Avoid to call update_slurm_reservation when it is not required
Previously we were calling scontrol command to update the reservation
every 10 minutes, but this is useless.

The only case where we need to update the reservation is to update the list
of nodes, this can happen only at cluster update time, so when Clustermgtd is restarted.
To identify this moment we are using the `_capacity_blocks_update_time` attribute.

Defined a new `_is_initialized` method and using it for `_update_slurm_reservation`

Signed-off-by: Enrico Usai <usai@amazon.com>