Update Capacity Blocks at every clustermgtd loop if there were errors updating Slurm reservations
Previously, in case of internal error during the creation/update/deletion of slurm reservations,
the code was trying to do the same action only after 10 minutes.

Now we're monitoring the `_slurm_reservation_update_errors` attribute
and if it has a value `> 0` we're saying that it's time to update.

This improves the responsiveness of the system.

1. When CB passes from inactive to active --> we need to remove the slurm reservation ASAP,
retrying many times in case of issues. Have the reservation removed is really important.

2. When CB passes from active to inactive --> if there are jobs associated to the nodes
and the nodes are not yet terminated it's not possible to create the reservation,
so we need to retry ASAP and not after 10 minutes.

Signed-off-by: Enrico Usai <usai@amazon.com>