Change log level to warning when temporarily partial EC2 info
Change log level to warning when temporarily unable to retrieve EC2 instance info from DescribeInstances call.

The from_describe_instance_data log a warning and throw an exception.
The from_describe_instance_data is called by
1. fleet_manager.launch_ec2_instances, but here the instance info are already populated in the _launch_instances, which calls the _get_instances_info that does have the retries, by calling _retrieve_instances_info_from_ec2
2. fleet_manager._retrieve_instances_info_from_ec2 that catch the exception

If at the end of the retires there are still instances info with partial data, in the _launch_instances these instances are finally logged as error

Signed-off-by: Luca Carrogu <carrogu@amazon.com>